requirements.txt
cffi==1.17.1
mpmath==1.3.0
numpy==2.3.1
pycparser==2.22
pycryptodome==3.23.0
sympy==1.14.0
